Ciudades
Se crearon las Pages de Ciudades, se implemento el buscador de ciudad. Falta hacer la parte de recibir la lista de ciudades

+++ app/src/main/java/com/example/appclima/presentacion/clima/ClimaEstado.kt
@@ -1,14 +1,15 @@
 package com.example.appclima.presentacion.clima
 
+import com.example.appclima.repository.modelos.Clima2
+
 sealed class ClimaEstado{
     data class Exitoso(
         val ciudad: String =  "",
-        val temperatura: Int = 0 ,
+        val temperatura: Double = 0.0 ,
         val descripcion : String ="",
-        val st : Int = 0,
-        val longitud :  Long=0,
-        val latitud : Long =0,
-
+        val st : Double = 0.0,
+        val longitud :  Double=0.0,
+        val latitud : Double =0.0,
         ): ClimaEstado()
     data class Error(
         val mensaje: String = ""
